[change] ファイルを分割
rooms.js, users.jsも参照
